Reorganize queue tests
